Добавлено условие при заполнении количества комнат
